requirements.txt
fastapi
uvicorn
pymupdf
openai
pinecone-client
python-docx
nltk
python-multipart
cohere
python-dotenv
streamlit
streamlit-modal